extract delete dialog to generic component
